pyproject.toml
[project]
name = "syntesa-lab-attendance-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "curl-cffi>=0.14.0",
    "fastapi[standard]>=0.128.6",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.11.0",
    "sqlmodel>=0.0.32",
    "tqdm>=4.67.3",
    "typer>=0.21.1",
]
